+++ test/rocket-utils/rocket-set-storage-scenarios.js
